Divide "core" and "client" modules.
