+++ .gitignore
@@ -5,3 +5,4 @@
 *.iml
 /tmp
 /store
+/*/target/
