+++ log-server-core/src/com/github/winmain/logserver/core/LogServer.scala
@@ -1,0 +1,11 @@
+package com.github.winmain.logserver.core
+
+import java.nio.charset.{Charset, StandardCharsets}
+
+object LogServer {
+
+  val StorageVersion = 2
+
+  val Charset: Charset = StandardCharsets.UTF_8
+
+}
