+++ src/com/github/winmain/logserver/core/storage/Storage.scala
@@ -4,10 +4,8 @@ import java.io.IOException
 import java.nio.charset.StandardCharsets
 import java.util
 
-import com.fasterxml.jackson.annotation.JsonValue
+import com.github.winmain.logserver.core.RecordId
 import com.github.winmain.logserver.utils.MurmurHash3
-
-import scala.util.Try
 
 object Storage {
   val Charset = StandardCharsets.UTF_8
@@ -20,68 +18,6 @@ object Storage {
     */
   val MaxBytesBuffer = 1024 * 1024
   //  val MagicEnding = 0xe0ffe0ff
-
-  sealed trait RecordId {
-    def hash: Int
-    def length: Int
-
-    @JsonValue
-    def jsonValue: Any =
-      this match {
-        case i: IntRecordId    => i.value
-        case s: StringRecordId => new String(s.value, Storage.Charset)
-        case EmptyRecordId     => null
-      }
-
-    override def equals(other: Any): Boolean = {
-      this match {
-        case i: IntRecordId =>
-          other.isInstanceOf[IntRecordId] && other
-            .asInstanceOf[IntRecordId]
-            .value == i.value
-        case s: StringRecordId =>
-          other.isInstanceOf[StringRecordId] && util.Arrays
-            .equals(other.asInstanceOf[StringRecordId].value, s.value)
-        case EmptyRecordId =>
-          this == EmptyRecordId
-      }
-    }
-  }
-
-  object RecordId {
-    val EmptyIdMarker: Byte = 0
-    val IntIdMarker: Byte = 1
-    val StringRecordMarker: Byte = 2
-
-    def apply(i: Int): RecordId = new IntRecordId(i)
-    def apply(s: String): RecordId =
-      new StringRecordId(s.getBytes(Storage.Charset))
-
-    def str(bytes: Array[Byte]): RecordId =
-      new StringRecordId(bytes)
-
-    def empty: RecordId = EmptyRecordId
-
-    def parse(s: String): RecordId =
-      Try(s.toInt).map(RecordId(_)).getOrElse(RecordId(s))
-  }
-
-  class IntRecordId(val value: Int) extends RecordId {
-    def hash: Int = value
-    def length: Int = 5
-    override def toString: String = value.toString
-  }
-
-  class StringRecordId(val value: Array[Byte]) extends RecordId {
-    def hash: Int = value.foldLeft(0)(_ + _ * 31)
-    def length: Int = 1 + 4 + value.length
-    override def toString: String = new String(value, Storage.Charset)
-  }
-
-  case object EmptyRecordId extends RecordId {
-    override def hash: Int = 0
-    override def length: Int = 1
-  }
 
   case class Header(timestamp: Long,
                     offset: Int,
